+++ htof/test/test_parse.py
@@ -7,9 +7,10 @@ import tempfile
 from ast import literal_eval
 
 from astropy.table import Table
-from htof.parse import HipparcosOriginalData, HipparcosRereductionData,\
-    GaiaData, DataParser, GaiaDR2, DecimalYearData
-from htof.parse import calculate_covariance_matrices, _match_filename_to_star_id
+from htof.parse import HipparcosOriginalData, HipparcosRereductionDVDBook,\
+    GaiaData, DataParser, GaiaDR2, GaiaeDR3, DecimalYearData, HipparcosRereductionJavaTool, digits_only, \
+    match_filename, get_nparam
+from htof.parse import calc_inverse_covariance_matrices
 
 
 class TestHipparcosOriginalData:
@@ -43,6 +44,10 @@ class TestHipparcosOriginalData:
         assert np.isclose(np.sin(data.scan_angle[5]), 0.7364, atol=.001)
         assert np.isclose(data.along_scan_errs[5], 2.0814, atol=.0001)
         assert np.isclose(data.residuals[5], 1.1021, atol=.0001)
+        data.parse(star_id=hip_id,
+                   intermediate_data_directory=test_data_directory,
+                   data_choice='BOTH')
+        assert len(data) == 66
 
     @pytest.mark.integration
     def test_concatenation(self, hip_id='027321'):
@@ -94,16 +99,15 @@ class TestHipparcosOriginalData:
         assert len(data) == 34
 
 
-class TestHipparcosRereductionData:
+class TestHipparcosRereductionDVDBook:
     def test_error_inflation_factor(self):
         u = 0.875291 # D. Michalik et al. 2014 Q factor for Hip 27321, calculated by hand
-        assert np.isclose(HipparcosRereductionData.error_inflation_factor(111, 5, -1.81), u, rtol=1e-5)
+        assert np.isclose(HipparcosRereductionDVDBook.error_inflation_factor(111, 5, -1.81), u, rtol=1e-5)
 
     def test_parse(self):
         test_data_directory = os.path.join(os.getcwd(), 'htof/test/data_for_tests/Hip2')
-        data = HipparcosRereductionData()
-        data.parse(star_id='027321',
-                   intermediate_data_directory=test_data_directory, convert_to_jd=False)
+        data = HipparcosRereductionDVDBook()
+        data.parse(star_id='027321', intermediate_data_directory=test_data_directory)
         nu = 111 - 5
         Q = nu * (np.sqrt(2/(9*nu))*-1.81 + 1 - 2/(9*nu))**3  # D. Michalik et al. 2014 Q factor for Hip 27321
         # Note that F2 = -1.81 in the CD intermediate data, while F2 = -1.63 on Vizier.
@@ -115,33 +119,98 @@ class TestHipparcosRereductionData:
         assert np.isclose(data._epoch[84], 1991.952)
         assert np.isclose(np.sin(data.scan_angle[84]), -0.8083, rtol=.01)
 
+    def test_reject_obs(self):
+        test_data_directory = os.path.join(os.getcwd(), 'htof/test/data_for_tests/Hip2')
+        data = HipparcosRereductionDVDBook()
+        data.parse(star_id='27321', intermediate_data_directory=test_data_directory)
+        assert len(data) == 111 - 0
+        assert data.rejected_epochs == []
+
+        data.parse(star_id='84', intermediate_data_directory=test_data_directory)
+        assert len(data) == 96 - 1
+        assert np.allclose(np.sort(data.rejected_epochs), [69])
+
+        data.parse(star_id='70', intermediate_data_directory=test_data_directory)
+        assert len(data) == 112 - 5
+        assert np.allclose(np.sort(data.rejected_epochs), [36, 55, 59, 64, 105])
+
+        data.parse(star_id='40', intermediate_data_directory=test_data_directory)
+        assert len(data) == 134 - 2
+        assert np.allclose(np.sort(data.rejected_epochs), [6, 53])
+
+        data.parse(star_id='072477', intermediate_data_directory=test_data_directory)
+        assert len(data) == 64 - 0
+        assert data.rejected_epochs == []
+
+
+class TestHipparcosRereductionJavaTool:
+    test_data_directory = os.path.join(os.getcwd(), 'htof/test/data_for_tests/Hip21')
+
+    def test_parse(self):
+        data = HipparcosRereductionJavaTool()
+        data.parse(star_id='27321', intermediate_data_directory=self.test_data_directory)
+        u = 1  # Error inflation factor
+        assert len(data) == 111
+        assert np.isclose(data._epoch[0], 1990.0055)
+        assert np.isclose(np.sin(data.scan_angle[0]), -0.9050, rtol=.01)
+        assert np.isclose(data.along_scan_errs.values[0], 0.80 * u)
+        assert np.isclose(data._epoch[84], 1991.9523)
+        assert np.isclose(np.sin(data.scan_angle[84]), -0.8083, rtol=.01)
+
+    def test_outlier_reject(self):
+        data = HipparcosRereductionJavaTool()
+        data.parse(star_id='27100', intermediate_data_directory=self.test_data_directory)
+        assert len(data) == 147 - 2  # num entries - num outliers
+        # outliers are marked with negative AL errors. Assert outliers are gone.
+        assert np.all(data.along_scan_errs > 0)
+
 
 class TestDataParser:
     def test_parse_raises_file_not_found_error(self):
         with pytest.raises(FileNotFoundError):
             test_data_directory = os.path.join(os.getcwd(), 'htof/test/data_for_tests/Hip2')
-            data = HipparcosRereductionData()
+            data = HipparcosRereductionDVDBook()
             data.parse(star_id='12gjas2',
-                       intermediate_data_directory=test_data_directory, convert_to_jd=False)
+                       intermediate_data_directory=test_data_directory)
 
     @mock.patch('htof.parse.glob.glob', return_value=['path/027321.dat', 'path/027321.dat'])
     def test_parse_raises_error_on_many_files_found(self, fake_glob):
         test_data_directory = os.path.join(os.getcwd(), 'htof/test/data_for_tests/Hip2')
-        data = HipparcosRereductionData()
-        with pytest.raises(ValueError):
+        data = HipparcosRereductionDVDBook()
+        with pytest.raises(FileNotFoundError):
             data.parse(star_id='027321',
-                       intermediate_data_directory=test_data_directory, convert_to_jd=False)
+                       intermediate_data_directory=test_data_directory)
 
     @mock.patch('htof.parse.pd.read_csv', return_value=None)
     @mock.patch('htof.parse.glob.glob', return_value=['path/127321.dat', 'path/27321.dat'])
-    def test_read_matches_filename_if_needed(self, fake_glob, fake_load):
-        test_data_directory = os.path.join(os.getcwd(), 'htof/test/data_for_tests/Hip2')
-        data = DataParser()
-        assert data.read_intermediate_data_file('27321', test_data_directory, None, None, None) is None
+    def test_read_on_couple_similar_filepaths(self, fake_glob, fake_load):
+        test_data_directory = os.path.join(os.getcwd(), 'path/')
+        DataParser().read_intermediate_data_file('27321', test_data_directory, None, None, None)
+        fake_load.assert_called_with('path/27321.dat', sep=None, skiprows=None, header=None, engine='python')
 
-    def test_match_filename_to_star_id(self):
-        paths = _match_filename_to_star_id('232', ['/fake/path/1232.dat', '/fake/path/23211.dat', '/fake/path/232.dat'])
-        assert paths == ['/fake/path/232.dat']
+    @mock.patch('htof.parse.pd.read_csv', return_value=None)
+    @mock.patch('htof.parse.glob.glob')
+    def test_read_on_many_filepaths(self, fake_glob, fake_load):
+        test_data_directory = os.path.join(os.getcwd(), 'path/')
+        fake_glob.return_value = ['/fake/path/1232.dat', '/fake/path/23211.dat', '/fake/path/232.dat']
+        DataParser().read_intermediate_data_file('232', test_data_directory, None, None, None)
+        fake_load.assert_called_with('/fake/path/232.dat', sep=None, skiprows=None, header=None, engine='python')
+        # Test hip2 style names
+        fake_glob.return_value = ['H075/HIP075290.d', 'H075/HIP075293.d', 'H107/HIP107529.d', 'H007/HIP007529.d', 'H117/HIP117529.d']
+        DataParser().read_intermediate_data_file('7529', test_data_directory, None, None, None)
+        fake_load.assert_called_with('H007/HIP007529.d', sep=None, skiprows=None, header=None, engine='python')
+
+    def test_match_filename(self):
+        paths = ['/hip/24374.txt', 'hip/43749.txt', 'hip/43740.txt', 'hip/4374.txt']
+        assert 'hip/4374.txt' == match_filename(paths, '4374')[0]
+
+    @mock.patch('htof.parse.pd.read_csv', return_value=None)
+    @mock.patch('htof.parse.glob.glob')
+    def test_read_on_irregular_paths(self, fake_glob, fake_load):
+        test_data_directory = os.path.join(os.getcwd(), 'path/')
+        fake_glob.return_value = ['/fake/path/SSADSx1232.dat', '/fake/path/WDASxs23211.dat', '/fake/path/OMGH232.dat']
+        DataParser().read_intermediate_data_file('232', test_data_directory, None, None, None)
+        fake_load.assert_called_with('/fake/path/OMGH232.dat', sep=None, skiprows=None, header=None, engine='python')
 
     def test_len(self):
         assert len(DataParser()) == 0
@@ -181,11 +250,11 @@ def test_trim_gaia_data():
     assert np.allclose(data.values.flatten(), [datemin, datemax])
 
 
-@mock.patch('htof.parse.calculate_covariance_matrices', return_value=np.array([np.ones((2, 2))]))
+@mock.patch('htof.parse.calc_inverse_covariance_matrices', return_value=np.array([np.ones((2, 2))]))
 def test_calculate_inverse_covariances(mock_cov_matrix):
     parser = DataParser()
     parser.calculate_inverse_covariance_matrices()
-    assert np.allclose(parser.inverse_covariance_matrix[0], 1/4 * np.ones((2, 2)))
+    assert np.allclose(parser.inverse_covariance_matrix[0], np.ones((2, 2)))
 
 
 class TestGaiaData:
@@ -205,6 +274,7 @@ class TestGaiaData:
     def test_parse_selects_valid_epochs(self):
         test_data_directory = os.path.join(os.getcwd(), 'htof/test/data_for_tests/GaiaDR2/IntermediateData')
         data = GaiaDR2(max_epoch=2458426.7784441218, min_epoch=2457143.4935643710)
+        data.DEAD_TIME_TABLE_NAME = None  # Do not reject dead times for this test
         data.parse(intermediate_data_directory=test_data_directory,
                    star_id='049699')
 
@@ -222,6 +292,32 @@ class TestGaiaData:
             assert np.isclose(d._epoch[0], 2456951.7659301492)
             assert np.isclose(d.scan_angle[0], -1.8904696884345342)
         assert len(data) == len(data2)
+
+    @pytest.mark.integration
+    def test_DR2parse_removes_dead_times(self):
+        # tests the dead time table against a fake source which has the first observation fall in a data gap
+        # and has the second observation not fall into any data gap.
+        test_data_directory = os.path.join(os.getcwd(), 'htof/test/data_for_tests/GaiaDR2/IntermediateData')
+        data = GaiaDR2(max_epoch=np.inf, min_epoch=-np.inf)
+        data.parse(intermediate_data_directory=test_data_directory,
+                   star_id='gaiatestsource01')
+
+        assert len(data._epoch) == 1
+        assert np.isclose(data._epoch.iloc[0], 2456893.28785)
+        assert np.isclose(data.scan_angle.iloc[0], -1.7804696884345342)
+
+    @pytest.mark.integration
+    def test_eDR3parse_removes_dead_times(self):
+        # tests the dead time table against a fake source which has the first observation fall in a data gap
+        # and has the second observation not fall into any data gap.
+        test_data_directory = os.path.join(os.getcwd(), 'htof/test/data_for_tests/GaiaeDR3/IntermediateData')
+        data = GaiaeDR3(max_epoch=np.inf, min_epoch=-np.inf)
+        data.parse(intermediate_data_directory=test_data_directory,
+                   star_id='gaiatestsource01')
+
+        assert len(data._epoch) == 1
+        assert np.isclose(data._epoch.iloc[0], 2456893.28785)
+        assert np.isclose(data.scan_angle.iloc[0], -1.7804696884345342)
 
 
 def test_write_with_missing_info():
@@ -257,8 +353,12 @@ def test_write():
 
 def test_calculating_covariance_matrices():
     scan_angles = pd.DataFrame(data=np.linspace(0, 2 * np.pi, 5))
-    covariances = calculate_covariance_matrices(scan_angles, cross_scan_along_scan_var_ratio=10)
-    assert len(covariances) == len(scan_angles)
+    icovariances = calc_inverse_covariance_matrices(scan_angles, cross_scan_along_scan_var_ratio=10)
+    assert len(icovariances) == len(scan_angles)
+    # calculate the covariance matrices so that we can compute the scan angle from the matrix.
+    covariances = np.zeros_like(icovariances)
+    for i in range(len(icovariances)):
+        covariances[i] = np.linalg.pinv(icovariances[i])
     assert np.allclose(covariances[-1], covariances[0])  # check 2pi is equivalent to 0.
     assert np.allclose(covariances[0], np.array([[10, 0], [0, 1]]))  # angle of 0 has AL parallel with DEC.
     for cov_matrix, scan_angle in zip(covariances, scan_angles.values.flatten()):
@@ -320,3 +420,13 @@ def angle_of_short_axis_of_error_ellipse(cov_matrix):
     x, y = vecs[:, 0]
     theta = np.arctan2(y, x) - np.pi/2
     return theta
+
+
+def test_digits_only():
+    assert '987978098098098' == digits_only("sdkjh987978asd098as0980a98sd")
+
+
+def test_get_nparam():
+    assert 5 == get_nparam('23115')
+    assert 5 == get_nparam('95')
+    assert 7 == get_nparam('97')